Consolidate hash build operators
Always use ParallelHashBuildOperator implementation.
Remove HashBuildOperator.
Rename ParallelHashBuildOperator to HashBuildOperator
